pyproject.toml
[tool.poetry]
name = "web-hw-12"
version = "0.1.0"
description = ""
authors = ["Max <maxkryvytskyiwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
alembic = "^1.13.0"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
uvicorn = "^0.24.0.post1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.5.2"}
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
psycopg2 = "^2.9.9"
 
 



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
